Split profile template service tests
Move shared profile/template fixtures and helper builders into a
dedicated fixture module, then separate series-profile and
episode-template service coverage into focused test modules.

Remove the deleted monolithic test module and its file-level
`too-many-lines` suppression.
