--- tests/test_profile_template_service.py
@@ -1,683 +1,0 @@
-"""Profile/template service tests.
-
-Summary
--------
-Unit and integration tests for canonical profile/template services.
-
-Purpose
--------
-Validate profile and template create/update flows, revision-conflict handling,
-history persistence, and structured brief generation.
-
-Usage
------
-Run this module through pytest as part of the canonical service suite.
-
-Example
--------
->>> pytest tests/test_profile_template_service.py -q
-"""
-
-# Profile/template service tests share revision and persistence fixtures.
-# pylint: disable=too-many-lines
-
-import dataclasses as dc
-import datetime as dt
-import itertools
-import typing as typ
-import uuid
-
-import pytest
-import pytest_asyncio
-
-from episodic.canonical import build_series_brief
-from episodic.canonical.domain import (
-    ApprovalState,
-    CanonicalEpisode,
-    EpisodeStatus,
-    ReferenceBinding,
-    ReferenceBindingTargetKind,
-    ReferenceDocument,
-    ReferenceDocumentKind,
-    ReferenceDocumentLifecycleState,
-    ReferenceDocumentRevision,
-    TeiHeader,
-)
-from episodic.canonical.profile_templates import (
-    AuditMetadata,
-    EpisodeTemplateData,
-    EpisodeTemplateUpdateFields,
-    RevisionConflictError,
-    SeriesProfileCreateData,
-    SeriesProfileUpdateFields,
-    UpdateEpisodeTemplateRequest,
-    UpdateSeriesProfileRequest,
-    create_episode_template,
-    create_series_profile,
-    list_history,
-    update_episode_template,
-    update_series_profile,
-)
-from episodic.canonical.storage import SqlAlchemyUnitOfWork
-
-if typ.TYPE_CHECKING:
-    from sqlalchemy.ext.asyncio import AsyncSession
-
-    from episodic.canonical.domain import (
-        EpisodeTemplate,
-        EpisodeTemplateHistoryEntry,
-        SeriesProfile,
-        SeriesProfileHistoryEntry,
-    )
-    from episodic.canonical.prompts import RenderedPrompt
-
-
-@dc.dataclass(slots=True)
-class BaseProfileFixture:
-    """Typed fixture payload for a base series profile."""
-
-    profile: SeriesProfile
-    profile_revision: int
-
-
-@dc.dataclass(slots=True)
-class BaseProfileWithTemplateFixture:
-    """Typed fixture payload for a base profile and one template."""
-
-    profile: SeriesProfile
-    profile_revision: int
-    template: EpisodeTemplate
-    template_revision: int
-
-
-@pytest_asyncio.fixture
-async def base_profile(
-    session_factory: typ.Callable[[], AsyncSession],
-) -> BaseProfileFixture:
-    """Create a reusable base series profile fixture."""
-    async with SqlAlchemyUnitOfWork(session_factory) as uow:
-        profile, profile_revision = await create_series_profile(
-            uow,
-            data=SeriesProfileCreateData(
-                slug="service-profile",
-                title="Service Profile",
-                description="Initial profile",
-                configuration={"tone": "neutral"},
-                guardrails={
-                    "instruction": "Keep the host voice calm and evidence-led.",
-                    "banned_phrases": ["smash that like button"],
-                },
-            ),
-            audit=AuditMetadata(
-                actor="author@example.com",
-                note="Initial version",
-            ),
-        )
-    return BaseProfileFixture(profile=profile, profile_revision=profile_revision)
-
-
-@pytest_asyncio.fixture
-async def base_profile_with_template(
-    session_factory: typ.Callable[[], AsyncSession],
-    base_profile: BaseProfileFixture,
-) -> BaseProfileWithTemplateFixture:
-    """Create a reusable base profile fixture with one episode template."""
-    profile = base_profile.profile
-    profile_revision = base_profile.profile_revision
-    async with SqlAlchemyUnitOfWork(session_factory) as uow:
-        template, template_revision = await create_episode_template(
-            uow,
-            series_profile_id=profile.id,
-            data=EpisodeTemplateData(
-                slug="weekly-template",
-                title="Weekly Template",
-                description="Template for weekly episodes",
-                structure={"segments": ["intro", "news", "outro"]},
-                guardrails={
-                    "instruction": "Open with the top story and close with a recap.",
-                    "required_sections": ["intro", "news", "outro"],
-                },
-            ),
-            audit=AuditMetadata(
-                actor="editor@example.com",
-                note="Initial template",
-            ),
-        )
-    return BaseProfileWithTemplateFixture(
-        profile=profile,
-        profile_revision=profile_revision,
-        template=template,
-        template_revision=template_revision,
-    )
-
-
-def _reconstruct_prompt_text(
-    *,
-    static_parts: tuple[str, ...],
-    interpolation_values: tuple[str, ...],
-) -> str:
-    """Rebuild rendered prompt text from its static/interpolation metadata."""
-    return "".join(
-        f"{part}{value}"
-        for part, value in itertools.zip_longest(
-            static_parts,
-            interpolation_values,
-            fillvalue="",
-        )
-    )
-
-
-def _assert_rendered_prompt_properties(
-    rendered_prompt: object,
-    *,
-    expected_expressions: list[str],
-) -> None:
-    """Verify RenderedPrompt contains expected interpolation and static metadata.
-
-    Args:
-        rendered_prompt: The RenderedPrompt object to verify.
-        expected_expressions: List of interpolation expressions that must be present.
-    """
-    prompt = typ.cast("RenderedPrompt", rendered_prompt)
-
-    assert prompt.interpolations, (
-        "Expected prompt rendering to include interpolation metadata."
-    )
-
-    for expression in expected_expressions:
-        assert any(item.expression == expression for item in prompt.interpolations), (
-            f"Expected interpolation metadata to include {expression}."
-        )
-
-    assert prompt.static_parts, (
-        "Expected prompt rendering to include static template parts."
-    )
-
-    assert (
-        _reconstruct_prompt_text(
-            static_parts=prompt.static_parts,
-            interpolation_values=tuple(item.value for item in prompt.interpolations),
-        )
-        == prompt.text
-    ), "Expected static and interpolation metadata to reconstruct text."
-
-
-def _build_brief_test_episode(
-    *,
-    profile_id: uuid.UUID,
-    episode_id: uuid.UUID,
-    now: dt.datetime,
-) -> tuple[CanonicalEpisode, TeiHeader]:
-    """Create an episode/header pair for episode-aware brief tests."""
-    episode = CanonicalEpisode(
-        id=episode_id,
-        series_profile_id=profile_id,
-        tei_header_id=uuid.uuid4(),
-        title="Episode-aware brief",
-        tei_xml="<TEI/>",
-        status=EpisodeStatus.DRAFT,
-        approval_state=ApprovalState.DRAFT,
-        created_at=now,
-        updated_at=now,
-    )
-    header = TeiHeader(
-        id=episode.tei_header_id,
-        title=episode.title,
-        payload={"file_desc": {"title": episode.title}},
-        raw_xml="<teiHeader/>",
-        created_at=now,
-        updated_at=now,
-    )
-    return episode, header
-
-
-def _build_cross_series_resolved_binding_fixture(
-    *,
-    now: dt.datetime,
-    foreign_profile_id: uuid.UUID,
-    target_profile_id: uuid.UUID,
-) -> tuple[ReferenceDocument, ReferenceDocumentRevision, ReferenceBinding]:
-    """Create a foreign-owned series binding for episode-aware brief tests."""
-    foreign_document = ReferenceDocument(
-        id=uuid.uuid4(),
-        owner_series_profile_id=foreign_profile_id,
-        kind=ReferenceDocumentKind.GUEST_PROFILE,
-        lifecycle_state=ReferenceDocumentLifecycleState.ACTIVE,
-        metadata={"name": "Foreign Guest"},
-        created_at=now,
-        updated_at=now,
-    )
-    foreign_revision = ReferenceDocumentRevision(
-        id=uuid.uuid4(),
-        reference_document_id=foreign_document.id,
-        content={"bio": "Foreign profile content"},
-        content_hash="foreign-hash-episode-aware",
-        author="author@example.com",
-        change_note="Cross-series revision for episode-aware brief",
-        created_at=now,
-    )
-    cross_series_binding = ReferenceBinding(
-        id=uuid.uuid4(),
-        reference_document_revision_id=foreign_revision.id,
-        target_kind=ReferenceBindingTargetKind.SERIES_PROFILE,
-        series_profile_id=target_profile_id,
-        episode_template_id=None,
-        ingestion_job_id=None,
-        effective_from_episode_id=None,
-        created_at=now,
-    )
-    return foreign_document, foreign_revision, cross_series_binding
-
-
-async def _seed_cross_series_resolved_binding_brief_scenario(
-    *,
-    uow: SqlAlchemyUnitOfWork,
-    profile_id: uuid.UUID,
-    now: dt.datetime,
-) -> uuid.UUID:
-    """Persist an episode plus a foreign-owned resolved binding for brief tests."""
-    episode_id = uuid.uuid4()
-    episode, header = _build_brief_test_episode(
-        profile_id=profile_id,
-        episode_id=episode_id,
-        now=now,
-    )
-    await uow.tei_headers.add(header)
-    await uow.flush()
-    await uow.episodes.add(episode)
-
-    foreign_profile, _ = await create_series_profile(
-        uow,
-        data=SeriesProfileCreateData(
-            slug="foreign-profile-episode-aware",
-            title="Foreign Profile Episode-aware",
-            description="Cross-series owner",
-            configuration={"tone": "direct"},
-            guardrails={},
-        ),
-        audit=AuditMetadata(
-            actor="author@example.com",
-            note="Create foreign profile for episode-aware brief",
-        ),
-    )
-    foreign_document, foreign_revision, cross_series_binding = (
-        _build_cross_series_resolved_binding_fixture(
-            now=now,
-            foreign_profile_id=foreign_profile.id,
-            target_profile_id=profile_id,
-        )
-    )
-    await uow.reference_documents.add(foreign_document)
-    await uow.reference_document_revisions.add(foreign_revision)
-    await uow.flush()
-    await uow.reference_bindings.add(cross_series_binding)
-    await uow.commit()
-    return episode_id
-
-
-class TestSeriesProfileService:
-    """Tests for series-profile service behaviour."""
-
-    @pytest.mark.asyncio
-    async def test_create_series_profile_creates_initial_history(
-        self,
-        session_factory: typ.Callable[[], AsyncSession],
-        base_profile: BaseProfileFixture,
-    ) -> None:
-        """Creating a profile also creates revision 1 history."""
-        profile = base_profile.profile
-        revision = base_profile.profile_revision
-
-        assert revision == 1, "Expected initial revision to be 1."
-
-        async with SqlAlchemyUnitOfWork(session_factory) as uow:
-            history = await list_history(
-                uow,
-                parent_id=profile.id,
-                kind="series_profile",
-            )
-        typed_history = typ.cast("list[SeriesProfileHistoryEntry]", history)
-        history = sorted(typed_history, key=lambda entry: entry.revision)
-
-        assert len(history) == 1, "Expected one profile history record."
-        first_entry = typ.cast("SeriesProfileHistoryEntry", history[0])
-        assert first_entry.revision == 1, "Expected first revision number to be 1."
-        assert first_entry.actor == "author@example.com", "Expected actor in history."
-
-    @pytest.mark.asyncio
-    async def test_update_series_profile_rejects_revision_conflicts(
-        self,
-        session_factory: typ.Callable[[], AsyncSession],
-        base_profile: BaseProfileFixture,
-    ) -> None:
-        """Updating with stale expected revision raises conflict."""
-        profile = base_profile.profile
-
-        async with SqlAlchemyUnitOfWork(session_factory) as uow:
-            with pytest.raises(RevisionConflictError, match=r"revision conflict"):
-                await update_series_profile(
-                    uow,
-                    request=UpdateSeriesProfileRequest(
-                        profile_id=profile.id,
-                        expected_revision=5,
-                        data=SeriesProfileUpdateFields(
-                            title="Profile Conflict Updated",
-                            description="Changed profile",
-                            configuration={"tone": "assertive"},
-                            guardrails={"instruction": "Stay factual."},
-                        ),
-                        audit=AuditMetadata(
-                            actor="editor@example.com",
-                            note="Conflict attempt",
-                        ),
-                    ),
-                )
-
-    @pytest.mark.asyncio
-    async def test_update_series_profile_updates_entity_and_appends_history(
-        self,
-        session_factory: typ.Callable[[], AsyncSession],
-        base_profile: BaseProfileFixture,
-    ) -> None:
-        """Updating with the current revision mutates entity data and history."""
-        profile = base_profile.profile
-        audit = AuditMetadata(
-            actor="reviewer@example.com",
-            note="Apply canonical profile update",
-        )
-        update_payload = SeriesProfileUpdateFields(
-            title="Service Profile Updated",
-            description="Updated profile description",
-            configuration={"tone": "assertive"},
-            guardrails={
-                "instruction": "Stay precise and cite uncertainty explicitly.",
-            },
-        )
-
-        async with SqlAlchemyUnitOfWork(session_factory) as uow:
-            updated_profile, updated_revision = await update_series_profile(
-                uow,
-                request=UpdateSeriesProfileRequest(
-                    profile_id=profile.id,
-                    expected_revision=1,
-                    data=update_payload,
-                    audit=audit,
-                ),
-            )
-
-        assert updated_revision == 2, "Expected profile revision to increment to 2."
-        assert updated_profile.title == update_payload.title, (
-            "Expected updated profile title."
-        )
-        assert updated_profile.description == update_payload.description, (
-            "Expected updated profile description."
-        )
-        assert updated_profile.configuration == update_payload.configuration, (
-            "Expected updated profile configuration."
-        )
-        assert updated_profile.guardrails == {
-            "instruction": "Stay precise and cite uncertainty explicitly.",
-            "banned_phrases": ["smash that like button"],
-        }, "Expected partial profile guardrail updates to preserve existing keys."
-
-        async with SqlAlchemyUnitOfWork(session_factory) as uow:
-            persisted_profile = await uow.series_profiles.get(profile.id)
-
-        assert persisted_profile is not None, "Expected updated profile to persist."
-        assert persisted_profile.guardrails == updated_profile.guardrails, (
-            "Expected persisted profile guardrails to match the merged update."
-        )
-
-        async with SqlAlchemyUnitOfWork(session_factory) as uow:
-            history = await list_history(
-                uow,
-                parent_id=profile.id,
-                kind="series_profile",
-            )
-        typed_history = typ.cast("list[SeriesProfileHistoryEntry]", history)
-        sorted_history = sorted(typed_history, key=lambda entry: entry.revision)
-
-        assert len(sorted_history) == 2, "Expected two profile history records."
-        latest_entry = sorted_history[-1]
-        assert latest_entry.revision == 2, "Expected latest revision number to be 2."
-        assert latest_entry.actor == audit.actor, "Expected actor in history."
-
-
-class TestEpisodeTemplateService:
-    """Tests for episode-template service behaviour."""
-
-    @pytest.mark.asyncio
-    async def test_update_episode_template_revision_conflict_raises(
-        self,
-        session_factory: typ.Callable[[], AsyncSession],
-        base_profile_with_template: BaseProfileWithTemplateFixture,
-    ) -> None:
-        """Updating a template with a stale revision raises conflict."""
-        template = base_profile_with_template.template
-        current_revision = base_profile_with_template.template_revision
-
-        stale_revision = current_revision - 1 if current_revision > 1 else 0
-
-        async with SqlAlchemyUnitOfWork(session_factory) as uow:
-            with pytest.raises(RevisionConflictError, match=r"revision conflict"):
-                await update_episode_template(
-                    uow,
-                    request=UpdateEpisodeTemplateRequest(
-                        template_id=template.id,
-                        expected_revision=stale_revision,
-                        data=EpisodeTemplateUpdateFields(
-                            title="Stale Template Update",
-                            description="Should fail",
-                            structure={"segments": ["intro", "outro"]},
-                            guardrails={"instruction": "Keep it brief."},
-                        ),
-                        audit=AuditMetadata(
-                            actor="editor@example.com",
-                            note="Stale update attempt",
-                        ),
-                    ),
-                )
-
-        async with SqlAlchemyUnitOfWork(session_factory) as uow:
-            history = await list_history(
-                uow,
-                parent_id=template.id,
-                kind="episode_template",
-            )
-        typed_history = typ.cast("list[EpisodeTemplateHistoryEntry]", history)
-        history = sorted(typed_history, key=lambda entry: entry.revision)
-
-        assert len(history) == 1, (
-            "Conflicting template update must not create a new history entry."
-        )
-        first_entry = typ.cast("EpisodeTemplateHistoryEntry", history[0])
-        assert first_entry.revision == current_revision, (
-            "History revision must remain at the last successful revision."
-        )
-
-    @pytest.mark.asyncio
-    async def test_create_episode_template_creates_history_and_brief(
-        self,
-        session_factory: typ.Callable[[], AsyncSession],
-        base_profile_with_template: BaseProfileWithTemplateFixture,
-    ) -> None:
-        """Creating a template records history and is retrievable in brief output."""
-        profile = base_profile_with_template.profile
-        template = base_profile_with_template.template
-        template_revision = base_profile_with_template.template_revision
-
-        assert template_revision == 1, "Expected initial template revision to be 1."
-
-        async with SqlAlchemyUnitOfWork(session_factory) as uow:
-            template_history = await list_history(
-                uow,
-                parent_id=template.id,
-                kind="episode_template",
-            )
-            template_history = sorted(
-                typ.cast("list[EpisodeTemplateHistoryEntry]", template_history),
-                key=lambda entry: entry.revision,
-            )
-            brief = await build_series_brief(
-                uow,
-                profile_id=profile.id,
-                template_id=template.id,
-            )
-
-        assert len(template_history) == 1, "Expected one template history record."
-        first_entry = typ.cast("EpisodeTemplateHistoryEntry", template_history[0])
-        assert first_entry.revision == 1, "Expected first template revision."
-        series_profile = typ.cast("dict[str, object]", brief["series_profile"])
-        templates = typ.cast("list[dict[str, object]]", brief["episode_templates"])
-        assert series_profile["id"] == str(profile.id), (
-            "Expected profile in structured brief."
-        )
-        assert series_profile["guardrails"] == profile.guardrails, (
-            "Expected profile guardrails in structured brief."
-        )
-        assert any(item["id"] == str(template.id) for item in templates), (
-            "Expected template in structured brief."
-        )
-        matched_template = next(
-            (item for item in templates if item["id"] == str(template.id)),
-            None,
-        )
-        assert matched_template is not None, "Expected template in structured brief."
-        assert matched_template["guardrails"] == template.guardrails, (
-            "Expected template guardrails in structured brief."
-        )
-
-    @pytest.mark.asyncio
-    async def test_update_episode_template_preserves_existing_guardrails(
-        self,
-        session_factory: typ.Callable[[], AsyncSession],
-        base_profile_with_template: BaseProfileWithTemplateFixture,
-    ) -> None:
-        """Partial guardrail updates must merge with persisted template rules."""
-        template = base_profile_with_template.template
-
-        async with SqlAlchemyUnitOfWork(session_factory) as uow:
-            updated_template, updated_revision = await update_episode_template(
-                uow,
-                request=UpdateEpisodeTemplateRequest(
-                    template_id=template.id,
-                    expected_revision=base_profile_with_template.template_revision,
-                    data=EpisodeTemplateUpdateFields(
-                        title="Weekly Template Updated",
-                        description="Updated template description",
-                        structure={"segments": ["intro", "analysis", "outro"]},
-                        guardrails={"instruction": "Close with a sourced takeaway."},
-                    ),
-                    audit=AuditMetadata(
-                        actor="editor@example.com",
-                        note="Merge template guardrails",
-                    ),
-                ),
-            )
-
-        assert updated_revision == 2, "Expected template revision to increment to 2."
-        assert updated_template.guardrails == {
-            "instruction": "Close with a sourced takeaway.",
-            "required_sections": ["intro", "news", "outro"],
-        }, "Expected partial template guardrail updates to preserve existing keys."
-
-    @pytest.mark.asyncio
-    async def test_build_series_brief_rejects_cross_series_reference_documents(
-        self,
-        session_factory: typ.Callable[[], AsyncSession],
-        base_profile_with_template: BaseProfileWithTemplateFixture,
-    ) -> None:
-        """Brief generation should reject references owned by another series."""
-        profile = base_profile_with_template.profile
-        template = base_profile_with_template.template
-        now = dt.datetime.now(dt.UTC)
-
-        async with SqlAlchemyUnitOfWork(session_factory) as uow:
-            foreign_profile, _ = await create_series_profile(
-                uow,
-                data=SeriesProfileCreateData(
-                    slug="foreign-profile",
-                    title="Foreign Profile",
-                    description="Cross-series owner",
-                    configuration={"tone": "direct"},
-                    guardrails={},
-                ),
-                audit=AuditMetadata(
-                    actor="author@example.com",
-                    note="Create foreign profile",
-                ),
-            )
-            foreign_document = ReferenceDocument(
-                id=uuid.uuid4(),
-                owner_series_profile_id=foreign_profile.id,
-                kind=ReferenceDocumentKind.GUEST_PROFILE,
-                lifecycle_state=ReferenceDocumentLifecycleState.ACTIVE,
-                metadata={"name": "Foreign Guest"},
-                created_at=now,
-                updated_at=now,
-            )
-            foreign_revision = ReferenceDocumentRevision(
-                id=uuid.uuid4(),
-                reference_document_id=foreign_document.id,
-                content={"bio": "Foreign profile content"},
-                content_hash="foreign-hash",
-                author="author@example.com",
-                change_note="Cross-series revision",
-                created_at=now,
-            )
-            cross_series_binding = ReferenceBinding(
-                id=uuid.uuid4(),
-                reference_document_revision_id=foreign_revision.id,
-                target_kind=ReferenceBindingTargetKind.EPISODE_TEMPLATE,
-                series_profile_id=None,
-                episode_template_id=template.id,
-                ingestion_job_id=None,
-                effective_from_episode_id=None,
-                created_at=now,
-            )
-            await uow.reference_documents.add(foreign_document)
-            await uow.reference_document_revisions.add(foreign_revision)
-            await uow.flush()
-            await uow.reference_bindings.add(cross_series_binding)
-            await uow.commit()
-
-        async with SqlAlchemyUnitOfWork(session_factory) as uow:
-            with pytest.raises(
-                ValueError,
-                match="does not belong to requested series profile",
-            ):
-                await build_series_brief(
-                    uow,
-                    profile_id=profile.id,
-                    template_id=template.id,
-                )
-
-    @pytest.mark.asyncio
-    async def test_build_series_brief_rejects_cross_series_resolved_bindings(
-        self,
-        session_factory: typ.Callable[[], AsyncSession],
-        base_profile_with_template: BaseProfileWithTemplateFixture,
-    ) -> None:
-        """Episode-aware brief generation should reject foreign resolved bindings."""
-        profile = base_profile_with_template.profile
-        template = base_profile_with_template.template
-        now = dt.datetime.now(dt.UTC)
-
-        async with SqlAlchemyUnitOfWork(session_factory) as uow:
-            episode_id = await _seed_cross_series_resolved_binding_brief_scenario(
-                uow=uow,
-                profile_id=profile.id,
-                now=now,
-            )
-
-        async with SqlAlchemyUnitOfWork(session_factory) as uow:
-            with pytest.raises(
-                ValueError,
-                match="does not belong to requested series profile",
-            ):
-                await build_series_brief(
-                    uow,
-                    profile_id=profile.id,
-                    template_id=template.id,
-                    episode_id=episode_id,
-                )
